+++ index.html
@@ -9,7 +9,7 @@
     <meta name="title" content="PLSS Monument Review : Provided by the UGRC" />
     <meta
       name="description"
-      content="A web mapping application that allows PLSS administrators to review and approve submissions."
+      content="A web mapping application that allows PLSS administrators to review and approve monument submissions."
     />
     <link rel="canonical" href="https://plss-review.ugrc.utah.gov/" />
     <!-- Open Graph / Facebook -->
@@ -18,7 +18,7 @@
     <meta property="og:title" content="PLSS Monument Review : Provided by the UGRC" />
     <meta
       property="og:description"
-      content="A web mapping application that will show you Discover base maps, kitchen sink react components, and the esri javascript api. It is a GitHub template that you can use to create your own website that is mobile friendly with dark mode."
+      content="A web mapping application that allows PLSS administrators to review and approve monument submissions."
     />
     <meta property="og:image" content="https://plss-review.ugrc.utah.gov/plss-social-card.png" />
     <meta property="og:image:width" content="1200" />
@@ -29,7 +29,7 @@
     <meta property="twitter:title" content="PLSS Monument Review : Provided by the UGRC" />
     <meta
       property="twitter:description"
-      content="A web mapping application that will show you Discover base maps, kitchen sink react components, and the esri javascript api. It is a GitHub template that you can use to create your own website that is mobile friendly with dark mode."
+      content="A web mapping application that allows PLSS administrators to review and approve monument submissions."
     />
     <meta property="twitter:image" content="https://plss-review.ugrc.utah.gov/plss-social-card.png" />
     <!-- Preloads -->
