+++ src/components/MapContainer.tsx
@@ -5,7 +5,7 @@ import { utahMercatorExtent } from '@ugrc/utilities/hooks';
 import { useEffect, useRef, useState } from 'react';
 import { useMap } from './hooks';
 
-export const MapContainer = ({ onClick }: { onClick?: __esri.ViewImmediateClickEventHandler }) => {
+export const MapContainer = ({ onClick }: { onClick?: __esri.ViewClickEventHandler }) => {
   const mapNode = useRef<HTMLDivElement | null>(null);
   const mapComponent = useRef<EsriMap | null>(null);
   const mapView = useRef<MapView>(null);
@@ -51,7 +51,7 @@ export const MapContainer = ({ onClick }: { onClick?: __esri.ViewImmediateClickE
   // add click event handlers
   useEffect(() => {
     if (onClick) {
-      clickHandler.current = mapView.current!.on('immediate-click', onClick);
+      clickHandler.current = mapView.current!.on('click', onClick);
     }
 
     return () => {
