+++ src/components/RejectionReasons.tsx
@@ -1,16 +1,7 @@
 import { Radio, RadioGroup, TextArea } from '@ugrc/utah-design-system';
 import { Controller, type Control } from 'react-hook-form';
+import type { FormValues } from './shared/types';
 
-export type FormValues = {
-  reason:
-    | 'missing-photo'
-    | 'sheet-incomplete'
-    | 'illegible-scan'
-    | 'incomplete-description'
-    | 'incomplete-sheet'
-    | 'other';
-  notes: string;
-};
 export function RejectionReasons({ control }: { control: Control<FormValues> }) {
   return (
     <form className="grid grid-cols-1 gap-4">
@@ -27,7 +18,7 @@ export function RejectionReasons({ control }: { control: Control<FormValues> }) 
             {...field}
           >
             <Radio value="missing-photo">Irrelevant or missing photos</Radio>
-            <Radio value="sheet-incomplete">Inaccurate or incomplete location information</Radio>
+            <Radio value="incomplete-location">Inaccurate or incomplete location information</Radio>
             <Radio value="illegible-scan">Illegible or poorly scanned document</Radio>
             <Radio value="incomplete-description">Incorrect monument description</Radio>
             <Radio value="incomplete-sheet">Missing required fields</Radio>
