+++ src/components/contexts/MapProvider.tsx
@@ -1,9 +1,8 @@
-import Graphic from '@arcgis/core/Graphic';
 import MapView from '@arcgis/core/views/MapView';
 import { useGraphicManager } from '@ugrc/utilities/hooks';
 import { createContext, type ReactNode, useState } from 'react';
+import type { GraphicOptions } from '../shared/types';
 
-type GraphicOptions = Graphic | Graphic[] | null;
 export const MapContext = createContext<{
   mapView: MapView | null;
   setMapView: (mapView: MapView | null) => void;
